pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "logicahome"
version = "0.3.0"
description = "Local-first MCP server and CLI for smart home control. Plug any AI into your home."
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "Andre Ambrosio" }]
requires-python = ">=3.11"
keywords = ["mcp", "smart-home", "home-automation", "tuya", "home-assistant", "matter", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
]

dependencies = [
    "mcp>=1.0.0,<2.0.0",
    "typer>=0.12.0,<1.0.0",
    "rich>=13.7.0",
    "pydantic>=2.6.0,<3.0.0",
    "aiohttp>=3.9.0,<4.0.0",
    "aiosqlite>=0.20.0",
    "platformdirs>=4.2.0",
    "pyyaml>=6.0",
    "uvicorn>=0.30.0",
    "starlette>=0.37.0",
    "zeroconf>=0.130.0",
]

[project.optional-dependencies]
tuya = ["tinytuya>=1.13.0"]
home-assistant = []  # uses aiohttp only
hue = []  # uses aiohttp only
shelly = []  # uses aiohttp only
esphome = ["aioesphomeapi>=24.0.0"]
matter = ["python-matter-server>=6.0.0"]
all = [
    "tinytuya>=1.13.0",
    "aioesphomeapi>=24.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
]

[project.scripts]
logicahome = "logicahome.cli:app"

[project.urls]
Homepage = "https://github.com/Rovemark/logicahome"
Issues = "https://github.com/Rovemark/logicahome/issues"
Documentation = "https://github.com/Rovemark/logicahome#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/logicahome"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "RET", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
